+++ scripts/lib/dependency-versions/index.mjs
@@ -1,0 +1,19 @@
+export {TRACKED_MODULES, createEmptyVersionSets} from './tracked-modules.mjs';
+export {
+  collectVersionsFromCorpus,
+  compareVersionsDesc,
+  normalizeVersion,
+  parseVersionsToml,
+  versionFromPatternMatch,
+  versionSetsToSortedRecords,
+} from './version-utils.mjs';
+export {collectAppVersionsFromProject, findGradleFiles} from './gradle-scan.mjs';
+export {collectAppVersionsFromApk} from './apk-scan.mjs';
+export {mapMetaInfVersionBaseToModule, mergeMetaInfEmbeddedVersions} from './apk-meta-inf.mjs';
+export {
+  mergeKotlinMetadataVersionsFromDex,
+  parseKotlinMetadataVersionsFromDexdump,
+} from './apk-kotlin-metadata.mjs';
+export {buildComparisonReport, compareModuleVersions} from './comparison.mjs';
+export {formatReport, mergeEspressoBuildConfigSuggestions} from './report.mjs';
+export {loadEspressoServerVersions} from './server-versions.mjs';
