+++ scripts/lib/diagnose/checks/index.mjs
@@ -1,0 +1,8 @@
+export {checkPrecompileInputKind} from './input-kind.mjs';
+export {checkInternetPermission} from './internet-permission.mjs';
+export {checkObfuscation} from './obfuscation.mjs';
+export {checkAndroidX} from './androidx.mjs';
+export {checkInitializationProvider} from './startup-provider.mjs';
+export {checkCompileSdk} from './compile-sdk.mjs';
+export {checkLifecycleExtensionsPin} from './lifecycle-extensions.mjs';
+export {mapDependencyChecks} from './dependency.mjs';
